+++ workspace/swiss-army-knife/src/bit_set/bit_set_aware.rs
@@ -19,7 +19,7 @@ macro_rules! bit_set_aware
 				{
 					Err(BitSetAwareTryFromU16Error::default())
 				}
-				else if unlikely!(value >= Self::LinuxMaximum)
+				else if unlikely!((value as u32) >= Self::LinuxMaximum)
 				{
 					Err(BitSetAwareTryFromU16Error::default())
 				}
@@ -62,7 +62,7 @@ macro_rules! bit_set_aware
 				{
 					Err(ParseNumberError::TooSmall)
 				}
-				else if unlikely!(value >= Self::LinuxMaximum)
+				else if unlikely!((value as u32) >= Self::LinuxMaximum)
 				{
 					Err(ParseNumberError::TooLarge)
 				}
